moved api, controllers, connect to db, create app, auth and models
