+++ root/create_app.py
@@ -1,0 +1,5 @@
+from flask import Flask
+from flask_cors import CORS
+
+app = Flask(__name__)
+CORS(app)
