+++ root/db/location/location_description_controller.py
