+++ root/db/qr/main_qr.py
