+++ root/db/text/translate_model.py
