+++ root/seo_optimisation_controller.py
@@ -32,3 +32,5 @@ class SeoOptimisationController:
 
 
 seoOptimisationController = SeoOptimisationController()
+
+
